fix(chain): give the tokens search the same row as the chains search
SearchChainToken was a stale copy of SearchChain: it missed the
extension row adopted in #4753, so the tokens search opened without a
Cancel button, without the pill field, and with a clear control that
rendered as a dim disc. The shared control now lives in
ExpandableSearch, driven by query/onQueryChange, and both searches
are thin wrappers that bind it to their own state provider.

Closes #4926
